refactor: place procurement semantics under domains
