+++ docs/domains/procurement/semantic-model.md
