+++ docs/project/handoff.md
@@ -16,7 +16,7 @@ Start with [AGENTS.md](../../AGENTS.md), then read the relevant [GitHub Issue](h
 
 ## Current architecture
 
-The system preserves evidence from synthetic/semi-structured procurement documents through structured mapping, source assertions, entity mentions, resolution decisions, canonicalized assertions, reconciliation, operational state, and derived intelligence. Core semantics remain framework-independent; adapters own external mechanics. Postgres is the intended canonical store, while lexical, vector, and graph systems are replaceable projections. Deterministic services compute and enforce policy; AI-assisted interfaces explain and route work. The accepted horizontal platform contract separates stage definitions, domain bindings/policy, and runtime implementation selection; it is documented in [ADR-022](../adr/022-domain-semantics-and-physical-stage-planning.md) and the [conformance contract](../architecture/domain-package-conformance.md), but is not implemented on `main`. See [architecture overview](../architecture/overview.md), [semantic model](../domain/semantic-model.md), and [architecture tradeoffs](../architecture.md).
+The system preserves evidence from synthetic/semi-structured procurement documents through structured mapping, source assertions, entity mentions, resolution decisions, canonicalized assertions, reconciliation, operational state, and derived intelligence. Core semantics remain framework-independent; adapters own external mechanics. Postgres is the intended canonical store, while lexical, vector, and graph systems are replaceable projections. Deterministic services compute and enforce policy; AI-assisted interfaces explain and route work. The accepted horizontal platform contract separates stage definitions, domain bindings/policy, and runtime implementation selection; it is documented in [ADR-022](../adr/022-domain-semantics-and-physical-stage-planning.md) and the [conformance contract](../architecture/domain-package-conformance.md), but is not implemented on `main`. Vertical-owned procurement semantics now live under `src/procurement_intelligence_lab/domains/procurement/`; horizontal package/compiler contracts remain under `src/procurement_intelligence_lab/domain/`. See [architecture overview](../architecture/overview.md), [semantic model](../domains/procurement/semantic-model.md), and [architecture tradeoffs](../architecture.md).
 
 ## Recently completed
