+++ src/procurement_intelligence_lab/application/review.py
@@ -7,8 +7,8 @@ from procurement_intelligence_lab.application.evidence_service import (
     EvidenceBackedClaim,
     inspect_bom_claims,
 )
-from procurement_intelligence_lab.domain.bom import Bom
-from procurement_intelligence_lab.domain.scope import Permission, RequestContext
+from procurement_intelligence_lab.domains.procurement.bom import Bom
+from procurement_intelligence_lab.domains.procurement.scope import Permission, RequestContext
 
 
 class ReviewReason(StrEnum):
