+++ src/procurement_intelligence_lab/domains/__init__.py
@@ -1,0 +1,1 @@
+"""Named domain verticals hosted by the platform."""
