+++ src/procurement_intelligence_lab/domains/procurement/__init__.py
@@ -1,0 +1,1 @@
+"""Procurement Intelligence vertical semantics and evidence contracts."""
