+++ src/procurement_intelligence_lab/domains/procurement/assertions.py
@@ -4,8 +4,8 @@ from dataclasses import dataclass
 from decimal import Decimal
 from enum import StrEnum
 
-from procurement_intelligence_lab.domain.bom import Bom, EvidenceRef
-from procurement_intelligence_lab.domain.identity import stable_id
+from procurement_intelligence_lab.domains.procurement.bom import Bom, EvidenceRef
+from procurement_intelligence_lab.domains.procurement.identity import stable_id
 
 
 class AssertionPredicate(StrEnum):
