+++ src/procurement_intelligence_lab/domains/procurement/bom.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 from decimal import Decimal
 from enum import StrEnum
 
-from procurement_intelligence_lab.domain.identity import stable_id
+from procurement_intelligence_lab.domains.procurement.identity import stable_id
 
 
 class EpistemicStatus(StrEnum):
