+++ src/procurement_intelligence_lab/domains/procurement/ledger.py
@@ -3,8 +3,8 @@
 from dataclasses import dataclass
 from datetime import datetime
 
-from procurement_intelligence_lab.domain.assertions import SourceAssertion
-from procurement_intelligence_lab.domain.identity import stable_id
+from procurement_intelligence_lab.domains.procurement.assertions import SourceAssertion
+from procurement_intelligence_lab.domains.procurement.identity import stable_id
 
 
 @dataclass(frozen=True)
