+++ src/procurement_intelligence_lab/domains/procurement/retrieval.py
@@ -4,9 +4,9 @@ from dataclasses import dataclass
 from datetime import datetime
 from enum import StrEnum
 
-from procurement_intelligence_lab.domain.identity import stable_id
-from procurement_intelligence_lab.domain.ledger import AssertionLedgerEntry
-from procurement_intelligence_lab.domain.scope import RequestContext
+from procurement_intelligence_lab.domains.procurement.identity import stable_id
+from procurement_intelligence_lab.domains.procurement.ledger import AssertionLedgerEntry
+from procurement_intelligence_lab.domains.procurement.scope import RequestContext
 
 
 class ProjectionKind(StrEnum):
