+++ src/procurement_intelligence_lab/domains/procurement/scope.py
