+++ src/procurement_intelligence_lab/ports/assertion_ledger.py
@@ -3,8 +3,8 @@
 from datetime import datetime
 from typing import Protocol
 
-from procurement_intelligence_lab.domain.assertions import SourceAssertion
-from procurement_intelligence_lab.domain.ledger import AssertionLedgerEntry
+from procurement_intelligence_lab.domains.procurement.assertions import SourceAssertion
+from procurement_intelligence_lab.domains.procurement.ledger import AssertionLedgerEntry
 
 
 class AssertionLedger(Protocol):
