+++ tests/contract/test_claim_semantics.py
@@ -6,8 +6,8 @@ from procurement_intelligence_lab.application.evidence_service import (
     ClaimKind,
     inspect_bom_claims,
 )
-from procurement_intelligence_lab.domain.bom import Bom, BomLine, EvidenceRef
-from procurement_intelligence_lab.domain.scope import Permission, RequestContext
+from procurement_intelligence_lab.domains.procurement.bom import Bom, BomLine, EvidenceRef
+from procurement_intelligence_lab.domains.procurement.scope import Permission, RequestContext
 
 
 def _context() -> RequestContext:
