+++ tests/unit/test_ledger.py
@@ -4,11 +4,11 @@ from decimal import Decimal
 import pytest
 
 from procurement_intelligence_lab.adapters.memory_ledger import InMemoryAssertionLedger
-from procurement_intelligence_lab.domain.assertions import (
+from procurement_intelligence_lab.domains.procurement.assertions import (
     AssertionPredicate,
     SourceAssertion,
 )
-from procurement_intelligence_lab.domain.bom import EvidenceRef
+from procurement_intelligence_lab.domains.procurement.bom import EvidenceRef
 
 
 def _assertion(quantity: str) -> SourceAssertion:
