+++ tests/unit/test_pipeline.py
@@ -1,9 +1,9 @@
 from decimal import Decimal
 
 from procurement_intelligence_lab.application.pipeline import run_bom_pipeline
-from procurement_intelligence_lab.domain.bom import Bom, BomLine, EvidenceRef
-from procurement_intelligence_lab.domain.evidence import EvidenceNodeKind
-from procurement_intelligence_lab.domain.resolution import ResolutionStatus
+from procurement_intelligence_lab.domains.procurement.bom import Bom, BomLine, EvidenceRef
+from procurement_intelligence_lab.domains.procurement.evidence import EvidenceNodeKind
+from procurement_intelligence_lab.domains.procurement.resolution import ResolutionStatus
 
 
 def test_bom_pipeline_preserves_layers_and_excludes_unresolved_state() -> None:
