+++ tests/unit/test_retrieval_projection.py
@@ -5,15 +5,18 @@ import pytest
 
 from procurement_intelligence_lab.adapters.memory_ledger import InMemoryAssertionLedger
 from procurement_intelligence_lab.adapters.memory_retrieval import InMemoryLexicalProjection
-from procurement_intelligence_lab.domain.assertions import AssertionPredicate, SourceAssertion
-from procurement_intelligence_lab.domain.bom import EvidenceRef
-from procurement_intelligence_lab.domain.ledger import AssertionLedgerEntry
-from procurement_intelligence_lab.domain.retrieval import (
+from procurement_intelligence_lab.domains.procurement.assertions import (
+    AssertionPredicate,
+    SourceAssertion,
+)
+from procurement_intelligence_lab.domains.procurement.bom import EvidenceRef
+from procurement_intelligence_lab.domains.procurement.ledger import AssertionLedgerEntry
+from procurement_intelligence_lab.domains.procurement.retrieval import (
     ProjectionBuildRequest,
     ProjectionKind,
     ProjectionStatus,
 )
-from procurement_intelligence_lab.domain.scope import (
+from procurement_intelligence_lab.domains.procurement.scope import (
     Permission,
     RequestContext,
     ScopeAuthorizationError,
