+++ tests/unit/test_review.py
@@ -2,7 +2,7 @@ from typing import cast
 
 import pytest
 
-from procurement_intelligence_lab.domain.scope import Permission, RequestContext
+from procurement_intelligence_lab.domains.procurement.scope import Permission, RequestContext
 from procurement_intelligence_lab.interfaces.web import (
     ReviewContextNotFoundError,
     claim_payload,
